fix lint

+++ src/instructions/extensions/m.rs
@@ -151,7 +151,7 @@ pub(crate) fn m() -> Vec<Instructor> {
         } else if dividend == i32::MIN || divisor == -1 {
           dividend as u64
         } else {
-          dividend.wrapping_div(divisor) as i32 as i64 as u64
+          dividend.wrapping_div(divisor) as i64 as u64
         };
         cpu.regs.set(rd, res);
         Ok(())
@@ -167,7 +167,7 @@ pub(crate) fn m() -> Vec<Instructor> {
         let dividend = cpu.regs[rs1] as u32;
         let divisor = cpu.regs[rs2] as u32;
         let res = if divisor == 0 {
-          u64::MAX as u64
+          u64::MAX
         } else {
           dividend.wrapping_div(divisor) as i32 as i64 as u64
         };
@@ -189,7 +189,7 @@ pub(crate) fn m() -> Vec<Instructor> {
         } else if dividend == i32::MIN || divisor == -1 {
           0
         } else {
-          dividend.wrapping_rem(divisor) as i32 as i64 as u64
+          dividend.wrapping_rem(divisor) as i64 as u64
         };
         cpu.regs.set(rd, res);
         Ok(())
